# Conflicts:
#	customerapp/src/main/java/com/tokopedia/tkpd/ConsumerRouterApplication.java
#	features/user/user_identification/src/main/AndroidManifest.xml
#	libraries/applink/src/main/java/com/tokopedia/applink/DeeplinkDFMapper.kt
#	libraries/applink/src/main/java/com/tokopedia/applink/DeeplinkMapper.kt
#	libraries/applink/src/main/java/com/tokopedia/applink/internal/ApplinkConstInternalGlobal.kt

+++ features/fintech/saldo_details/src/main/java/com/tokopedia/saldodetails/view/fragment/SaldoDepositFragment.java
@@ -36,6 +36,7 @@ import com.tokopedia.applink.internal.ApplinkConstInternalGlobal;
 import com.tokopedia.remoteconfig.FirebaseRemoteConfigImpl;
 import com.tokopedia.remoteconfig.RemoteConfig;
 import com.tokopedia.remoteconfig.RemoteConfigKey;
+import com.tokopedia.saldodetails.R;
 import com.tokopedia.saldodetails.activity.SaldoDepositActivity;
 import com.tokopedia.saldodetails.commom.analytics.SaldoDetailsConstants;
 import com.tokopedia.saldodetails.contract.SaldoDetailContract;
@@ -105,8 +106,8 @@ public class SaldoDepositFragment extends BaseDaggerFragment
     private boolean isSellerEnabled;
     private SaldoTransactionHistoryFragment saldoHistoryFragment;
 
-    private long sellerSaldoBalance;
-    private long buyerSaldoBalance;
+    private float sellerSaldoBalance;
+    private float buyerSaldoBalance;
     private float totalSaldoBalance;
     private LinearLayout saldoTypeLL;
     private LinearLayout merchantDetailLL;
@@ -516,12 +517,12 @@ public class SaldoDepositFragment extends BaseDaggerFragment
     }
 
     @Override
-    public long getSellerSaldoBalance() {
+    public float getSellerSaldoBalance() {
         return sellerSaldoBalance;
     }
 
     @Override
-    public long getBuyerSaldoBalance() {
+    public float getBuyerSaldoBalance() {
         return buyerSaldoBalance;
     }
 
@@ -677,13 +678,13 @@ public class SaldoDepositFragment extends BaseDaggerFragment
     }
 
     @Override
-    public void setBuyerSaldoBalance(long balance, String text) {
+    public void setBuyerSaldoBalance(float balance, String text) {
         buyerSaldoBalance = balance;
         buyerBalanceTV.setText(text);
     }
 
     @Override
-    public void setSellerSaldoBalance(long amount, String formattedAmount) {
+    public void setSellerSaldoBalance(float amount, String formattedAmount) {
         sellerSaldoBalance = amount;
         sellerBalanceTV.setText(formattedAmount);
     }
